extract replay config sheet to new component
